# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Reccomendation",
        "description": "Get All Reccomendation Books",
        "contact": {
            "email": "a.osamamahmoud96@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/intervals": {
            "get": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Get all intervals",
                "description": "Retrieves all intervals of all inserted intervals by users",
                "operationId": "IntervalsList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of the rows per the page [ pagination ]",
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the desired page number [ pagination ]",
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of intervals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_name": {
                                                "type": "string",
                                                "example": "Ahmed Osama"
                                            },
                                            "book_name": {
                                                "type": "string",
                                                "example": "Harry poter"
                                            },
                                            "start_page": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "end_page": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Add a new interval",
                "description": "Adds a new interval to the collection and returns the added interval details.\n     *      [ Message will be sended after submit]",
                "operationId": "storeInterval",
                "requestBody": {
                    "description": "Pass interval details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "start_page",
                                    "end_page"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "interger",
                                        "example": 10
                                    },
                                    "book_id": {
                                        "type": "interger",
                                        "example": 1
                                    },
                                    "start_page": {
                                        "type": "interger",
                                        "example": 1
                                    },
                                    "end_page": {
                                        "type": "interger",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interval successfully added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "interger",
                                            "example": 10
                                        },
                                        "book_id": {
                                            "type": "interger",
                                            "example": 1
                                        },
                                        "start_page": {
                                            "type": "interger",
                                            "example": 1
                                        },
                                        "end_page": {
                                            "type": "interger",
                                            "example": 15
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have entered this interval before"
                                        },
                                        "errors": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "You have entered this interval before"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/intervals/{id}": {
            "get": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Get a single Interval by ID",
                "description": "Retrieves the details of a single Interval given its ID.",
                "operationId": "getIntervalById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the inteval to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "book_id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "start_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "end_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Update a single Intervals's data",
                "description": "Updates the specified fields of a interval given its ID.",
                "operationId": "updateInterval",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the interval to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass interval details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "start_page",
                                    "end_page"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "interger",
                                        "example": 10
                                    },
                                    "book_id": {
                                        "type": "interger",
                                        "example": 1
                                    },
                                    "start_page": {
                                        "type": "interger",
                                        "example": 1
                                    },
                                    "end_page": {
                                        "type": "interger",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interval successfully added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "interger",
                                            "example": 10
                                        },
                                        "book_id": {
                                            "type": "interger",
                                            "example": 1
                                        },
                                        "start_page": {
                                            "type": "interger",
                                            "example": 1
                                        },
                                        "end_page": {
                                            "type": "interger",
                                            "example": 15
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have entered this interval before"
                                        },
                                        "errors": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "You have entered this interval before"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Delete a single interval",
                "description": "Deletes an interval from the collection given its ID.",
                "operationId": "deleteInterval",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the interval to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Delete operation"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
...
    }
}